+++ src/designer/elsa-workflows-studio/src/drivers/checkbox-driver.tsx
@@ -1,0 +1,15 @@
+﻿import {PropertyDisplayDriver} from "../services/property-display-driver";
+import {ActivityModel, ActivityPropertyDescriptor} from "../models";
+import {h, State} from "@stencil/core";
+import {getOrCreateProperty, setActivityModelProperty} from "../utils/utils";
+
+export class CheckboxDriver implements PropertyDisplayDriver {
+
+  display(activity: ActivityModel, property: ActivityPropertyDescriptor) {
+    const prop = getOrCreateProperty(activity, property.name);
+    return <elsa-checkbox-property propertyDescriptor={property} propertyModel={prop} />;
+  }
+
+  update(activity: ActivityModel, property: ActivityPropertyDescriptor, form: FormData) {
+  }
+}
