--- src/designer/elsa-workflows-studio/src/property-display-drivers/index.ts
@@ -1,6 +1,0 @@
-﻿export * from './null-property-driver';
-export * from './single-line-driver';
-export * from './multiline-driver';
-export * from './check-list-driver';
-export * from './checkbox-driver';
-export * from './dropdown-driver';
