--- src/designer/elsa-workflows-studio/src/property-display-drivers/multiline-driver.tsx
@@ -1,15 +1,0 @@
-﻿import {PropertyDisplayDriver} from "../services/property-display-driver";
-import {ActivityModel, ActivityPropertyDescriptor} from "../models";
-import {h} from "@stencil/core";
-import {getOrCreateProperty} from "../utils/utils";
-
-export class MultilineDriver implements PropertyDisplayDriver {
-
-  display(activity: ActivityModel, property: ActivityPropertyDescriptor) {
-    const prop = getOrCreateProperty(activity, property.name);
-    return <elsa-multi-line-property propertyDescriptor={property} propertyModel={prop}/>;
-  }
-
-  update(activity: ActivityModel, property: ActivityPropertyDescriptor, form: FormData) {
-  }
-}
